# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Joshua Tree Sales Rep App API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get current user profile
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateMe
      tags: [users]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/me/profile:
    patch:
      operationId: updateMyProfile
      tags: [users]
      summary: Update current user's profile (self-only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileBody"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /users/{userId}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get a specific user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update a user (admin)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete a user (admin)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Deals ───────────────────────────────────────────────────────
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        clerkId:
          type: string
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, rep]
        avatarUrl:
          type: ["string", "null"]
        accentColor:
          type: string
        hometown:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        hawaiiGoal:
          type: ["string", "null"]
        favoriteService:
          type: ["string", "null"]
        totalPoints:
          type: integer
        territoryId:
          type: ["integer", "null"]
        territoryName:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        level:
          type: integer
        nextLevelAt:
          type: integer
        pointsThisLevel:
          type: integer
        pointsPerLevel:
          type: integer
        currentStreak:
          type: integer
        bestStreak:
          type: integer
        streakAtRisk:
          type: boolean
      required:
        - id
        - clerkId
        - name
        - email
        - role
        - accentColor
        - totalPoints
        - createdAt
        - level
        - nextLevelAt
        - pointsThisLevel
        - pointsPerLevel
        - currentStreak
        - bestStreak
        - streakAtRisk
    UpdateProfileBody:
      type: object
      properties:
        name:
          type: string
        avatarUrl:
          type: ["string", "null"]
        accentColor:
          type: string
        hometown:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        hawaiiGoal:
          type: ["string", "null"]
        favoriteService:
          type: ["string", "null"]
...
    UpdateUserBody:
      type: object
      properties:
        name:
          type: string
        role:
          type: string
          enum: [admin, rep]
        avatarUrl:
          type: ["string", "null"]
        territoryId:
          type: ["integer", "null"]
        accentColor:
          type: ["string", "null"]
        hometown:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        hawaiiGoal:
          type: ["string", "null"]
        favoriteService:
          type: ["string", "null"]
...